+++ slack_bot/bot_home_view.py
